Fixed bug
Fixed bug where player could still access and talk to an NPC even after exhausting all their dialogue options. Game now automatically moves on to the next NPC with available dialogue options (Note: Prone to bugs in the future due to how the NPC scrolling works, be careful) and if all the NPCs in a building have been fully talked with, the building will no longer be highlighted and the player cannot click on the building anymore.